backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
PyMuPDF==1.23.8
python-multipart==0.0.6
gunicorn==21.2.0 